server/requirements.txt
Django==2.0.8
django-cors-headers==2.4.0
django-haystack==2.8.1
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
drf-haystack==1.8.3
drf-nested-routers==0.90.2
flake8==3.5.0
mccabe==0.6.1
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.3
python-dotenv==0.9.0
pytz==2018.5
six==1.11.0
Whoosh==2.7.4
